+++ buildpacks/php/tests/integration/utils.rs
@@ -1,3 +1,4 @@
+use fs_err as fs;
 use libcnb_test::{
     assert_contains, BuildConfig, BuildpackReference, ContainerConfig, TestContext, TestRunner,
 };
@@ -15,7 +16,7 @@ use std::time::Duration;
 ///a
 /// The smoke integration tests need to ensure the container runs as expected.
 /// This function is catering to that use-case and is not useful in other contexts.
-fn start_container_assert_basic_http_response(
+pub(crate) fn start_container_assert_basic_http_response(
     context: &TestContext,
     expected_http_response_body_contains: &str,
 ) {
@@ -116,4 +117,40 @@ pub(crate) fn default_buildpacks() -> Vec<BuildpackReference> {
         BuildpackReference::CurrentCrate,
         BuildpackReference::Other(String::from("heroku/procfile")),
     ]
+}
+
+/// Writes the given php version to the composer.json and composer.lock
+#[cfg(test)]
+pub(crate) fn set_php_version(app_dir: &Path, version_req: &str) {
+    let json_path = app_dir.join("composer.json");
+    let lock_path = app_dir.join("composer.lock");
+
+    let mut composer_json =
+        serde_json::from_str::<serde_json::Map<_, _>>(&fs::read_to_string(&json_path).unwrap())
+            .unwrap();
+    composer_json
+        .get_mut("require")
+        .unwrap()
+        .as_object_mut()
+        .unwrap()
+        .insert(
+            "php".to_string(),
+            serde_json::to_value(version_req).unwrap(),
+        );
+    fs::write(json_path, serde_json::to_string(&composer_json).unwrap()).unwrap();
+
+    let mut composer_lock = serde_json::from_str::<serde_json::Map<_, _>>(
+        &fs::read_to_string(app_dir.join("composer.lock")).unwrap(),
+    )
+    .unwrap();
+    composer_lock
+        .get_mut("platform")
+        .unwrap()
+        .as_object_mut()
+        .unwrap()
+        .insert(
+            "php".to_string(),
+            serde_json::to_value(version_req).unwrap(),
+        );
+    fs::write(lock_path, serde_json::to_string(&composer_lock).unwrap()).unwrap();
 }