+++ buildpacks/php/src/platform.rs
@@ -41,7 +41,7 @@ pub(crate) fn heroku_stack_name_for_target(target: &Target) -> Result<String, St
         ..
     } = target;
     match (os.as_str(), distro_name.as_str(), distro_version.as_str()) {
-        ("linux", "ubuntu", v @ ("20.04" | "22.04" | "24.04")) => {
+        ("linux", "ubuntu", v @ ("22.04" | "24.04")) => {
             Ok(format!("heroku-{}", v.strip_suffix(".04").unwrap_or(v)))
         }
         _ => Err(format!("{os}-{distro_name}-{distro_version}")),
@@ -62,7 +62,7 @@ pub(crate) fn platform_base_url_for_target(target: &Target) -> Url {
         let stack_name = heroku_stack_name_for_target(target)
             .expect("Internal error: could not determine Heroku stack name for OS/distro");
         match v {
-            "20.04" | "22.04" => stack_name,
+            "22.04" => stack_name,
             _ => format!("{stack_name}-{arch}"),
         }
     } else {