+++ buildpacks/php/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Drop support for heroku-20 ([#197](https://github.com/heroku/buildpacks-php/pull/197))
+
 ## [1.0.1] - 2025-04-29
 
 ### Fixed