+++ buildpacks/php/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2025-05-13
+
 ### Changed
 
 - Drop support for heroku-20 ([#197](https://github.com/heroku/buildpacks-php/pull/197))
@@ -83,7 +85,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation
 
-[unreleased]: https://github.com/heroku/buildpacks-php/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-php/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/heroku/buildpacks-php/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/heroku/buildpacks-php/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/heroku/buildpacks-php/compare/v0.2.4...v1.0.0
 [0.2.4]: https://github.com/heroku/buildpacks-php/compare/v0.2.3...v0.2.4